# ffs-api/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "description": "API for user authentication, creation, and management of schools, districts, bids, and tasks"
  },
  "paths": {
...
    "/api/schools/detailed/{districtId}": {
      "get": {
        "summary": "Get detailed schools by district ID",
        "description": "Retrieve detailed information about all schools in a specific district, including shipping addresses and contact information for display on the District Details page.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "The ID of the district to retrieve schools for",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schools retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchoolsDetailedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid district ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission to access this district",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - District not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{districtId}": {
      "post": {
        "summary": "Create a school",
        "description": "Create a new school within a specified district.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district where the school will be created",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchoolRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "School created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSchoolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid district ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get schools by district ID",
        "description": "Retrieve all schools belonging to a specific district identified by its ID.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district to retrieve schools from",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of schools in the district",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchoolResponseItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid district ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{districtId}/{id}/archive": {
      "put": {
        "summary": "Archive a school",
        "description": "Archive an existing school within a district, marking it as inactive.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district the school belongs to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the school to archive",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School archived successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{districtId}/{id}/activate": {
      "put": {
        "summary": "Activate a school",
        "description": "Activate an existing school within a district, marking it as active.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district the school belongs to",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the school to activate",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School activated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schools/{districtId}/{schoolId}/details": {
      "get": {
        "summary": "Get comprehensive school details",
        "description": "Retrieve comprehensive information about a specific school including district name, contact details, and billing information.",
        "tags": ["Schools"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the district",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "schoolId", 
            "in": "path",
            "required": true,
            "description": "Unique identifier of the school",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchoolDetailsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid district ID or school ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "School or district not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateSchoolResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the created school"
          },
          "name": {
            "type": "string",
            "description": "Name of the school"
          },
          "district_id": {
            "type": "integer",
            "description": "ID of the district the school belongs to"
          }
        },
        "required": ["id", "name", "district_id"]
      },
...
      "SchoolResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the school"
          },
          "name": {
            "type": "string",
            "description": "Name of the school"
          },
          "address": {
            "type": "string",
            "description": "Physical address of the school"
          },
          "city": {
            "type": "string",
            "description": "City where the school is located"
          },
          "state": {
            "type": "string",
            "description": "State where the school is located"
          },
          "zip_code": {
            "type": "string",
            "description": "Postal code of the school"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number of the school"
          },
          "enrollment": {
            "type": "integer",
            "description": "Number of students enrolled"
          },
          "district_id": {
            "type": "integer",
            "description": "ID of the district the school belongs to"
          }
        },
        "required": ["id", "name", "district_id"]
      },
      "GetSchoolsDetailedResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SchoolDetailedResponseItem"
        }
      },
...
    }
  }
}